feat: 修正部分語法錯誤
