+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "axios": "^1.6.0",
     "bootstrap": "^5.3.2",
+    "bootstrap-icons": "^1.13.1",
     "pinia": "^3.0.3",
     "vue": "^3.5.17",
     "vue-router": "^4.5.1"
