+++ src/components/common/ErrorMessage.vue
@@ -5,7 +5,7 @@
       <strong>錯誤：</strong>{{ message }}
       <button 
         v-if="showRetry" 
-        @click="$emit('retry')"
+        @click="emit('retry')"
         class="btn btn-sm btn-outline-danger ms-3"
       >
         重試
@@ -13,30 +13,30 @@
     </div>
     <button 
       v-if="dismissible"
-      @click="$emit('dismiss')"
+      @click="emit('dismiss')"
       type="button" 
       class="btn-close ms-auto"
     ></button>
   </div>
 </template>
 
-<script>
-export default {
-  name: 'ErrorMessage',
-  props: {
-    message: {
-      type: String,
-      required: true
-    },
-    showRetry: {
-      type: Boolean,
-      default: false
-    },
-    dismissible: {
-      type: Boolean,
-      default: true
-    }
+<script setup>
+// Props 定義 (JavaScript 版本)
+const props = defineProps({
+  message: {
+    type: String,
+    required: true
   },
-  emits: ['retry', 'dismiss']
-}
+  showRetry: {
+    type: Boolean,
+    default: false
+  },
+  dismissible: {
+    type: Boolean,
+    default: true
+  }
+})
+
+// Emits 定義
+const emit = defineEmits(['retry', 'dismiss'])
 </script>
