+++ src/components/common/LoadingSpinner.vue
@@ -5,41 +5,20 @@
   </div>
 </template>
 
-<script>
-export default {
-  name: 'LoadingSpinner',
-  props: {
-    size: {
-      type: String,
-      default: '20px'
-    },
-    text: {
-      type: String,
-      default: ''
-    },
-    inline: {
-      type: Boolean,
-      default: false
-    }
+<script setup>
+// Props 定義 (JavaScript 版本)
+const props = defineProps({
+  size: {
+    type: String,
+    default: '20px'
+  },
+  text: {
+    type: String,
+    default: ''
+  },
+  inline: {
+    type: Boolean,
+    default: false
   }
-}
+})
 </script>
-
-<style scoped>
-.loading-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 20px;
-}
-
-.loading-container.inline {
-  display: inline-flex;
-  padding: 0;
-}
-
-.loading-text {
-  color: #6c757d;
-  font-size: 0.9rem;
-}
-</style>
