+++ src/components/weather/SearchInput.vue
@@ -41,7 +41,7 @@
         <h6 class="suggestions-header">搜尋結果</h6>
         <div
           v-for="(suggestion, index) in suggestions"
-          :key="suggestion.name + suggestion.country"
+          :key="`${suggestion.name}-${suggestion.country}`"
           @click="selectSuggestion(suggestion)"
           class="suggestion-item"
           :class="{ active: selectedIndex === index }"
@@ -75,167 +75,144 @@
   </div>
 </template>
 
-<script>
-import { ref, computed, watch } from 'vue'
+<script setup>
+import { ref, computed } from 'vue'
 import { useWeatherStore } from '@/stores/weather'
 import { useHistoryStore } from '@/stores/history'
 import { useLocationStore } from '@/stores/location'
 import weatherAPI from '@/services/weatherApi'
 import LoadingSpinner from '@/components/common/LoadingSpinner.vue'
 
-export default {
-  name: 'SearchInput',
-  components: {
-    LoadingSpinner
-  },
-  setup() {
-    const weatherStore = useWeatherStore()
-    const historyStore = useHistoryStore()
-    const locationStore = useLocationStore()
+// Stores
+const weatherStore = useWeatherStore()
+const historyStore = useHistoryStore()
+const locationStore = useLocationStore()
 
-    const searchQuery = ref('')
-    const suggestions = ref([])
-    const showSuggestions = ref(false)
-    const selectedIndex = ref(-1)
-    const searchTimeout = ref(null)
-    const loading = ref(false)
+// Reactive data
+const searchQuery = ref('')
+const suggestions = ref([])
+const showSuggestions = ref(false)
+const selectedIndex = ref(-1)
+const searchTimeout = ref(null)
+const loading = ref(false)
 
-    const gettingLocation = computed(() => locationStore.isLocationLoading)
-    const recentSearches = computed(() => historyStore.recentSearches)
+// Computed
+const gettingLocation = computed(() => locationStore.isLocationLoading)
+const recentSearches = computed(() => historyStore.recentSearches)
 
-    // 防抖搜尋
-    const handleInput = () => {
-      if (searchTimeout.value) {
-        clearTimeout(searchTimeout.value)
-      }
-
-      selectedIndex.value = -1
-
-      if (searchQuery.value.trim().length >= 2) {
-        searchTimeout.value = setTimeout(async () => {
-          await searchCities()
-        }, 300)
-      } else {
-        suggestions.value = []
-      }
-    }
-
-    const searchCities = async () => {
-      try {
-        const results = await weatherAPI.searchCities(searchQuery.value.trim())
-        suggestions.value = results.slice(0, 5)
-      } catch (error) {
-        console.error('City search error:', error)
-        suggestions.value = []
-      }
-    }
-
-    const handleKeydown = (event) => {
-      const totalItems = suggestions.value.length + 
-        (searchQuery.value.trim() ? 0 : recentSearches.value.length)
-
-      switch (event.key) {
-        case 'ArrowDown':
-          event.preventDefault()
-          selectedIndex.value = Math.min(selectedIndex.value + 1, totalItems - 1)
-          break
-        case 'ArrowUp':
-          event.preventDefault()
-          selectedIndex.value = Math.max(selectedIndex.value - 1, -1)
-          break
-        case 'Enter':
-          event.preventDefault()
-          if (selectedIndex.value >= 0) {
-            selectByIndex()
-          } else {
-            handleSearch()
-          }
-          break
-        case 'Escape':
-          showSuggestions.value = false
-          selectedIndex.value = -1
-          break
-      }
-    }
-
-    const selectByIndex = () => {
-      if (selectedIndex.value < suggestions.value.length) {
-        selectSuggestion(suggestions.value[selectedIndex.value])
-      } else {
-        const historyIndex = selectedIndex.value - suggestions.value.length
-        selectHistory(recentSearches.value[historyIndex])
-      }
-    }
-
-    const selectSuggestion = (suggestion) => {
-      searchQuery.value = suggestion.name
-      showSuggestions.value = false
-      selectedIndex.value = -1
-      handleSearch()
-    }
-
-    const selectHistory = (item) => {
-      searchQuery.value = item.name
-      showSuggestions.value = false
-      selectedIndex.value = -1
-      handleSearch()
-    }
-
-    const removeHistory = (item) => {
-      historyStore.removeFromHistory(item.name)
-    }
-
-    const handleSearch = async () => {
-      const city = searchQuery.value.trim()
-      if (!city) return
-
-      loading.value = true
-      try {
-        await weatherStore.fetchWeatherByCity(city)
-        if (weatherStore.currentWeather) {
-          historyStore.addToHistory(city)
-          showSuggestions.value = false
-        }
-      } catch (error) {
-        console.error('Search error:', error)
-      } finally {
-        loading.value = false
-      }
-    }
-
-    const getCurrentLocation = async () => {
-      const success = await locationStore.getCurrentLocation()
-      if (success && locationStore.userLocation) {
-        const { lat, lon } = locationStore.userLocation
-        await weatherStore.fetchWeatherByCoords(lat, lon)
-      }
-    }
-
-    const handleBlur = () => {
-      // 延遲隱藏建議，讓點擊事件可以觸發
-      setTimeout(() => {
-        showSuggestions.value = false
-      }, 200)
-    }
-
-    return {
-      searchQuery,
-      suggestions,
-      showSuggestions,
-      selectedIndex,
-      loading,
-      gettingLocation,
-      recentSearches,
-      handleInput,
-      handleKeydown,
-      selectSuggestion,
-      selectHistory,
-      removeHistory,
-      handleSearch,
-      getCurrentLocation,
-      handleBlur
-    }
+// Methods
+const handleInput = () => {
+  if (searchTimeout.value) {
+    clearTimeout(searchTimeout.value)
   }
+
+  selectedIndex.value = -1
+
+  if (searchQuery.value.trim().length >= 2) {
+    searchTimeout.value = setTimeout(async () => {
+      await searchCities()
+    }, 300)
+  } else {
+    suggestions.value = []
+  }
+}
+
+const searchCities = async () => {
+  try {
+    const results = await weatherAPI.searchCities(searchQuery.value.trim())
+    suggestions.value = results.slice(0, 5)
+  } catch (error) {
+    console.error('City search error:', error)
+    suggestions.value = []
+  }
+}
+
+const handleKeydown = (event) => {
+  const totalItems = suggestions.value.length + 
+    (searchQuery.value.trim() ? 0 : recentSearches.value.length)
+
+  switch (event.key) {
+    case 'ArrowDown':
+      event.preventDefault()
+      selectedIndex.value = Math.min(selectedIndex.value + 1, totalItems - 1)
+      break
+    case 'ArrowUp':
+      event.preventDefault()
+      selectedIndex.value = Math.max(selectedIndex.value - 1, -1)
+      break
+    case 'Enter':
+      event.preventDefault()
+      if (selectedIndex.value >= 0) {
+        selectByIndex()
+      } else {
+        handleSearch()
+      }
+      break
+    case 'Escape':
+      showSuggestions.value = false
+      selectedIndex.value = -1
+      break
+  }
+}
+
+const selectByIndex = () => {
+  if (selectedIndex.value < suggestions.value.length) {
+    selectSuggestion(suggestions.value[selectedIndex.value])
+  } else {
+    const historyIndex = selectedIndex.value - suggestions.value.length
+    selectHistory(recentSearches.value[historyIndex])
+  }
+}
+
+const selectSuggestion = (suggestion) => {
+  searchQuery.value = suggestion.name
+  showSuggestions.value = false
+  selectedIndex.value = -1
+  handleSearch()
+}
+
+const selectHistory = (item) => {
+  searchQuery.value = item.name
+  showSuggestions.value = false
+  selectedIndex.value = -1
+  handleSearch()
+}
+
+const removeHistory = (item) => {
+  historyStore.removeFromHistory(item.name)
+}
+
+const handleSearch = async () => {
+  const city = searchQuery.value.trim()
+  if (!city) return
+
+  loading.value = true
+  try {
+    await weatherStore.fetchWeatherByCity(city)
+    if (weatherStore.currentWeather) {
+      historyStore.addToHistory(city)
+      showSuggestions.value = false
+    }
+  } catch (error) {
+    console.error('Search error:', error)
+  } finally {
+    loading.value = false
+  }
+}
+
+const getCurrentLocation = async () => {
+  const success = await locationStore.getCurrentLocation()
+  if (success && locationStore.userLocation) {
+    const { lat, lon } = locationStore.userLocation
+    await weatherStore.fetchWeatherByCoords(lat, lon)
+  }
+}
+
+const handleBlur = () => {
+  // 延遲隱藏建議，讓點擊事件可以觸發
+  setTimeout(() => {
+    showSuggestions.value = false
+  }, 200)
 }
 </script>
