+++ src/main.js
@@ -8,7 +8,7 @@ import router from './router'
 // Bootstrap CSS 和 JS
 import 'bootstrap/dist/css/bootstrap.min.css'
 import 'bootstrap/dist/js/bootstrap.bundle.min.js'
-
+import 'bootstrap-icons/font/bootstrap-icons.css'
 
 import './assets/styles/main.css'   // 統一風格樣式的地方
